# docs/api/openapi.yaml
components:
  schemas:
...
    api.SwaggerAssignRequest:
      properties:
        assignee_id:
          example: 874d6314-6353-45e9-ab2a-5fe930ea4dbc
          type: string
      type: object
...
    api.SwaggerErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/api.SwaggerErrorDetail'
      type: object
...
    api.SwaggerTicketResponse:
      properties:
        assignee_id:
          type: string
        created_at:
          type: string
        description:
          example: The login button does not work
          type: string
        due_at:
          type: string
        id:
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
          type: string
        labels:
          items:
            type: string
          type: array
          uniqueItems: false
        priority:
          example: 1
          type: integer
        rank:
          example: '0|aaaaaa:'
          type: string
        reporter_id:
          example: 874d6314-6353-45e9-ab2a-5fe930ea4dbc
          type: string
        resolved_at:
          type: string
        space_id:
          example: b2c3d4e5-f6a7-8901-bcde-f12345678901
          type: string
        status:
          example: open
          type: string
        title:
          example: Fix login button
          type: string
        updated_at:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 'JWT Bearer token. Format: "Bearer <token>". Obtain via POST /api/v1/auth/login'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: hello@azimuthalhq.com
    name: Azimuthal HQ
    url: https://azimuthalhq.com
  description: |-
    The fully open-source alternative to Jira, Confluence, and Jira Service Desk.
    Self-hostable, single binary, Apache 2.0 licensed.

    ## Authentication
    All endpoints except /auth/login and /auth/register require a Bearer JWT token.
    Obtain a token via POST /api/v1/auth/login, then include it as:
    `Authorization: Bearer <your-token>`
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Azimuthal API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /spaces/{spaceID}/tickets:
    get:
      description: Returns all tickets in the specified space.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.SwaggerTicketResponse'
                type: array
          description: List of tickets
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid space ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: List tickets
      tags:
      - tickets
    post:
      description: Creates a new ticket in the specified space. Reporter is set from
        the JWT.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerCreateTicketRequest'
                description: Ticket details
                summary: body
        description: Ticket details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerTicketResponse'
          description: Ticket created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Create ticket
      tags:
      - tickets
  /spaces/{spaceID}/tickets/{ticketID}:
    delete:
      description: Soft-deletes a ticket by ID.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Ticket ID (UUID)
        in: path
        name: ticketID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Delete ticket
      tags:
      - tickets
    get:
      description: Returns a single ticket by ID.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Ticket ID (UUID)
        in: path
        name: ticketID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerTicketResponse'
          description: Ticket details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Get ticket
      tags:
      - tickets
    patch:
      description: Updates an existing ticket's title, description, priority, and
        labels.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Ticket ID (UUID)
        in: path
        name: ticketID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerUpdateTicketRequest'
                description: Updated fields
                summary: body
        description: Updated fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerTicketResponse'
          description: Updated ticket
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Update ticket
      tags:
      - tickets
  /spaces/{spaceID}/tickets/{ticketID}/assign:
    delete:
      description: Removes the current assignee from a ticket.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Ticket ID (UUID)
        in: path
        name: ticketID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerTicketResponse'
          description: Updated ticket
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Unassign ticket
      tags:
      - tickets
    post:
      description: Assigns a ticket to a user by ID.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Ticket ID (UUID)
        in: path
        name: ticketID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerAssignRequest'
                description: Assignee
                summary: body
        description: Assignee
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerTicketResponse'
          description: Updated ticket
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Already assigned
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Assign ticket
      tags:
      - tickets
  /spaces/{spaceID}/tickets/{ticketID}/status:
    post:
      description: Changes the status of a ticket (e.g. open -> in_progress -> resolved
        -> closed).
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Ticket ID (UUID)
        in: path
        name: ticketID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerTransitionRequest'
                description: New status
                summary: body
        description: New status
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerTicketResponse'
          description: Updated ticket
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid status
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid transition
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Transition ticket status
      tags:
      - tickets
...
servers:
- url: localhost:8080/api/v1